+++ src/pages/Class/AllProblems/AllProblemDetail.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router-dom';
 
 import { SUB_PATH } from '@/constants';
+import { Problem } from '@/pages/Class/components';
 
-import { Problem } from '../Problem/components';
 import { useProblemQuery } from './hooks';
 
 export function AllProblemDetail() {
