+++ src/pages/Class/AllProblems/ProblemEdit.tsx
@@ -1,12 +1,12 @@
 import { useParams } from 'react-router-dom';
 
-import { ProblemForm } from './components';
-import { useEditProblemMutation } from './hooks';
-import { useProblemQuery } from '../AllProblems/hooks';
+import { ProblemForm } from '@/pages/Class/components';
+
+import { useEditProblemMutation, useProblemQuery } from './hooks';
 
 export function ProblemEdit() {
   const { problemId } = useParams() as { problemId: string };
-  /** FIXME: 파일이 존재하는 쿼리로 변경 */
+  /** FIXME: 파일이 존재하는 쿼리로 변경 + contest problem API 추가 */
   const { data } = useProblemQuery(problemId);
   const { mutate } = useEditProblemMutation(problemId);
