+++ src/pages/Class/AllProblems/api/index.ts
@@ -1,4 +1,4 @@
-import { api } from '@/api';
+import { api, fileApi } from '@/api';
 
 const API_URL = `/problems`;
 
@@ -10,12 +10,16 @@ const getProblem = (problemId: string) => {
   return api.get(`${API_URL}/${problemId}`);
 };
 
-const deleteProblem = (problemId: string) => {
-  return api.delete(`${API_URL}/${problemId}`);
+const editProblem = (problemId: string, payload: FormData) => {
+  return fileApi.put(`${API_URL}/${problemId}/`, payload);
 };
 
 const editProblemPublic = (problemId: string) => {
   return api.post(`${API_URL}/${problemId}/check/`);
 };
 
-export { getAllProblemList, getProblem, editProblemPublic, deleteProblem };
+const deleteProblem = (problemId: string) => {
+  return api.delete(`${API_URL}/${problemId}`);
+};
+
+export { getAllProblemList, getProblem, editProblem, editProblemPublic, deleteProblem };
