+++ src/pages/Class/AllProblems/hooks/query/index.ts
@@ -1,4 +1,5 @@
 export * from './useAllProblemListQuery';
 export * from './useProblemQuery';
+export * from './useEditProblemMutation';
 export * from './useEditProblemPublicMutation';
 export * from './useDeleteProblemMutation';
