+++ src/pages/Class/AllProblems/index.ts
@@ -1,2 +1,3 @@
 export * from './AllProblemList';
 export * from './AllProblemDetail';
+export * from './ProblemEdit';
