+++ src/pages/Class/Contest/ContestProblemDetail.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router-dom';
 
 import { SUB_PATH } from '@/constants';
+import { Problem } from '@/pages/Class/components';
 
-import { Problem } from '../Problem/components';
 import { useContestProblemQuery } from './hooks';
 
 export function ContestProblemDetail() {
