+++ src/pages/Class/Contest/hooks/query/useCreateContestProblemMutation.ts
@@ -2,13 +2,13 @@ import { useMutation, UseMutationOptions } from 'react-query';
 import { AxiosError, AxiosResponse } from 'axios';
 import { useNavigate } from 'react-router-dom';
 
-import { createProblem } from '../../api';
+import { createContestProblem } from '../../api';
 
-export const useCreateProblemMutation = (
+export const useCreateContestProblemMutation = (
   options?: UseMutationOptions<AxiosResponse, AxiosError, FormData>
 ) => {
   const navigate = useNavigate();
-  return useMutation((payload) => createProblem(payload), {
+  return useMutation((payload) => createContestProblem(payload), {
     ...options,
     onSuccess: () => {
       navigate(-1);
