+++ src/pages/Class/Contest/index.ts
@@ -1,7 +1,8 @@
 export * from './ClassContest';
-export * from './ClassContestProblemList';
 export * from './ClassContestListEdit';
+export * from './ClassContestProblemList';
 
+export * from './ContestProblemCreate';
 export * from './ContestProblemDetail';
 export * from './ContestProblemLeaderBoard';
 export * from './ContestProblemSubmission';
