--- src/pages/Class/Problem/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Problem';
-export * from './ProblemForm';
