--- src/pages/Class/Problem/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './query';
