--- src/pages/Class/Problem/hooks/query/index.ts
@@ -1,2 +1,0 @@
-export * from './useCreateProblemMutation';
-export * from './useEditProblemMutation';
