--- src/pages/Class/Problem/index.ts
@@ -1,5 +1,0 @@
-export * from './ProblemCreate';
-export * from './ProblemEdit';
-
-export * from './ProblemDescription';
-export * from './ProblemData';
