+++ src/pages/Class/components/index.ts
@@ -1,2 +1,6 @@
 export * from './ClassEditModal';
 export * from './ClassFormModal';
+export * from './Problem';
+export * from './ProblemForm';
+export * from './ProblemData';
+export * from './ProblemDescription';
